:recycle:Vueの本一覧部分を共通化
