+++ app/assets/stylesheets/application.css.scss
@@ -25,6 +25,9 @@
 .push-top{
   margin-top: 10px;
 }
+.push-bottom{
+  margin-bottom: 10px;
+}
 .push-side{
   margin-right: 10px;
   margin-left: 10px;
