+++ app/javascript/packs/components/searchedBooks.vue
