+++ app/views/api/authors/show.json.jbuilder
@@ -1,3 +1,7 @@
 # frozen_string_literal: true
 
-json.extract! @author, :id, :name, :birthday, :books
+json.set! :books do
+  json.array! @books do |book|
+    json.extract! book, :id, :title, :author_id, :rakuten_book_info
+  end
+end
