+++ app/views/books/search.html.slim
@@ -2,7 +2,6 @@
 
 #searchPage
   search
-
 = javascript_pack_tag 'search'
 
 sass:
